Move alpha headers up to lua/config, use headers in candy namespace
