+++ lua/free/config/alpha.lua
@@ -252,7 +252,7 @@ local buttons = {
 local header = {
   type = "text",
   -- From https://gist.github.com/sRavioli/d6fb0a813b6affc171976b7dd09764d3
-  val = require("free.config.alpha.headers")["random"],
+  val = require("config.alpha-headers")["random"],
   opts = {
     position = "center",
     hl = "AlphaHeader",
